Dictionaries for DECLARED variables (any type) parsed form the source program
Stores all declared parsed data types, that has to have quick access for its value. Declared, but not initialized or already evaluated vars has value "None" in the option. This way all variables are quicker to access in the situation of many occurances of variable in the same rule or expression, although it's not memory-friendly.