"Derivation" field parser
Derivation set mode in which the rules are considered at the start of the program. These may be executed form top to the bottom (still recursively calling each other), started only non-redundant ones (NOT IMPLEMENTED in final version) or the ones chosen by user. The last one is not really tested, but should ask the user for input numbers of the rules to be executed and do it one by one.