"Loop-over" variable field parser
Parses field for name of numeric variable  used for storing the number of loops of the program to be executed in this specific version of ExSys interpreter - so it's not native ExSys feature. At the start of the program the name of the value is checked and if is correct and stores correct value, the program can be started in the for loop, holding the values from the prevoious executions. This mode has beeen added to simulate (or testing) expert-system-based, fuzzy logic, control systems. Alternatively, it can be executed only once, for typical mode of operation.