"Or current" field parser
"y(es)" or "n(o)" values are parsed by this function and stored in order to execute rules in more custom mode. It's up to user whether or not the truthness of the qualifier value truthness should be logically summed with the previous one, already existing in the program memory. Default is none, so the new truthness of the value overwrites it, but ExSys has allowed to "or" it with the previous which may be useful for some programs.