Qualifier parser
Parses one occurence of the qualifier in the source file. Qualifier variable can be considered as enum in this expert system interpreter, but its values (enumerations) can be described by the certain level of truthness (depending on the mode). In fuzzy mode, the truthness is a float value between 0 and 1.