Calculates qualifier value truthness (in ExSys way) and assigns the truthness to value in dictionary
The truthness has to be calculated as  fuzzy product of firing level and value of truthness from assignment. It can be also summed in a fuzzy way with current value if it's specified in the basic attributes. The output of this process (called fuzzy implication) is assigned to specified value of the qualifier.