Section for parser of arithmetic expressions resulting in float
Section covers parsers for Expr type which is recursive type forming a tree from arithmetic operators and its arguments. The expression is stored in this form in a rule memory being evaluated in case of the rule firing (activation). The result can be used both in assignment ("THEN:" part of a rule) and comparison ("IF:" part of a rule).