Rule parser section
Consists of parsers of rule's elements. A rule has antecendent part (a "IF:" part) and consequent part (a "THEN:" part). The antecendent is parsed as a one expression (in recursive, tree structure) results in boolean value in the process of evaluation. Evaluation happens when the rule has assignment to one demanded variable in its consequent ("THEN:" part). In fuzzy inference, the important thing is firing (activation) level of the rule that takes part in calculating a truthness value.