Calculates fuzzy sum of specified values of qualifier
Takes a list of values (enums, string*float list) with membership levels from qualifier dictionary and list of values to be summed (string list). Requires correct name of the particular values otherwise sum with 0. Does not inferes qualifier values truthness because function is intended to work with the values directly from the dictionary.