Second qualifier definition - stores output information in fuzzy way
The value of this qulifier is calculated by rules using first qualifier as an input. The values (again, MAŁO, UMIARKOWANIE, DUŻO) has its fuzzified value of truthness that are used to calculate defuzzified output of the program. The output demanded numeric value is stored in MOC ("power") variable. The method used to defuzzification is hardcoded and it's method of heights.