Section for rules definitions
Rule consists of antecendent ("IF" part) and consequent ("THEN") . The first one represents a boolean condition that has to be met. In fuzzy systems, the antecendent results in some value between 0 and 1 that has to be combined with values from assignments from consequent in a process of fuzzy implication. Typically (and in this system) the method of implication is fuzzy product of rule and assignment value truthness.